+++ src/main/java/ru/sbt/mipt/oop/EventProcessor.java
@@ -1,0 +1,6 @@
+package ru.sbt.mipt.oop;
+
+public class EventProcessor {
+    public void runEvent(SensorEvent event, SmartHome smartHome) {
+    }
+}
