+++ src/main/java/ru/sbt/mipt/oop/LightsEventProcessor.java
@@ -1,0 +1,26 @@
+package ru.sbt.mipt.oop;
+
+import static ru.sbt.mipt.oop.SensorEventType.LIGHT_OFF;
+import static ru.sbt.mipt.oop.SensorEventType.LIGHT_ON;
+
+public class LightsEventProcessor extends EventProcessor {
+    @Override
+    public void runEvent(SensorEvent event, SmartHome smartHome) {
+        if (event.getType() == LIGHT_ON || event.getType() == LIGHT_OFF) {
+            // событие от источника света
+            for (Room room : smartHome.getRooms()) {
+                for (Light light : room.getLights()) {
+                    if (light.getId().equals(event.getObjectId())) {
+                        if (event.getType() == LIGHT_ON) {
+                            light.setOn(true);
+                            System.out.println("Light " + light.getId() + " in room " + room.getName() + " was turned on.");
+                        } else {
+                            light.setOn(false);
+                            System.out.println("Light " + light.getId() + " in room " + room.getName() + " was turned off.");
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
