+++ src/main/java/ru/sbt/mipt/oop/SmartHomeReader.java
@@ -1,0 +1,17 @@
+package ru.sbt.mipt.oop;
+
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import com.google.gson.Gson;
+
+public class SmartHomeReader {
+    public SmartHome ReadSmartHome() throws IOException {
+        // считываем состояние дома из файла
+        Gson gson = new Gson();
+        String json = new String(Files.readAllBytes(Paths.get("smart-home-1.js")));
+        return gson.fromJson(json, SmartHome.class);
+    }
+
+}
